Prototype1 Progress3
Populated the Prefabs folder created earlier with the food prefab and the animal prefabs.

Added the following scripts: ShootPrefab to clone the food prefab and shoot them forward; MoveForward to allow the food prefab and the animals to move forward (food moves up, animals move down); DetectCollisions to allow the food prefab to detect when it has collided with an animal and destroy both itself and the animal; DestroyOutOfBounds to detect when food or animals are out of bounds, destroying them.