Prototype2 Progress6
Added the HealthSystem with it's hearts and management over the player's health as animals run past the bottom bound of the screen. Imported the HealthSystem.cs file from BB, adjusted HealthSystem.cs (only a little, followed the video), DestroyOutOfBounds.cs (so that it can make calls to the TakeDamage() method), and SoawnManager.cs (to stop the game when the player has run out of health).